+++ exporters/otlp/otlpgrpc/example_test.go
@@ -26,6 +26,7 @@ import (
 	"go.opentelemetry.io/otel/exporters/otlp"
 	"go.opentelemetry.io/otel/exporters/otlp/otlpgrpc"
 	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/global"
 	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
 	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
@@ -187,7 +188,7 @@ func Example_withDifferentSignalCollectors() {
 		controller.WithPusher(exp),
 		controller.WithCollectPeriod(2*time.Second),
 	)
-	otel.SetMeterProvider(pusher.MeterProvider())
+	global.SetMeterProvider(pusher.MeterProvider())
 
 	if err := pusher.Start(ctx); err != nil {
 		log.Fatalf("could not start metric controoler: %v", err)
@@ -202,7 +203,7 @@ func Example_withDifferentSignalCollectors() {
 	}()
 
 	tracer := otel.Tracer("test-tracer")
-	meter := otel.Meter("test-meter")
+	meter := global.Meter("test-meter")
 
 	// Recorder metric example
 	counter := metric.Must(meter).
