+++ internal/global/benchmark_test.go
@@ -21,6 +21,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/label"
+	metricglobal "go.opentelemetry.io/otel/metric/global"
 )
 
 func BenchmarkGlobalInt64CounterAddNoSDK(b *testing.B) {
@@ -29,7 +30,7 @@ func BenchmarkGlobalInt64CounterAddNoSDK(b *testing.B) {
 	// global no-op system against a registered SDK.
 	global.ResetForTest()
 	ctx := context.Background()
-	sdk := otel.Meter("test")
+	sdk := metricglobal.Meter("test")
 	labs := []label.KeyValue{label.String("A", "B")}
 	cnt := Must(sdk).NewInt64Counter("int64.counter")
